Corrected noun Defenders -> Attackers
Under the heading ## Defender Mindset
After describing that Defenders are awash in lists, was:
"There's one problem with all of this. Defenders don't have a list of assets—they have a graph." Which contradicted what had already been said, so presumed should be Attackers in place of Defenders in this sentence.